fix some erros
