+++ WhiteboardFormatter/app/src/main/java/com/example/whiteboardformatter/data/dao/Dao.kt
@@ -2,30 +2,32 @@ package com.example.whiteboardformatter.data.dao
 
 import androidx.lifecycle.LiveData
 import androidx.room.*
-import com.example.yabamiru.data.model.Task
-import com.example.yabamiru.data.model.TaskAndTaskTags
+import com.example.whiteboardformatter.data.model.Text
+import com.example.whiteboardformatter.data.model.WhiteboardEntity
 
 @Dao
 interface TableDao{
 
-    @Query("SELECT * FROM whiteboard")
-    fun getAll():LiveData<List<Whiteboard>>
+    @Query("SELECT * FROM whiteboardentity")
+    fun getAll():LiveData<List<WhiteboardEntity>>
 
     @Query("SELECT * FROM Text WHERE whiteboardId=(:whiteboardId)")
     fun getById(whiteboardId:Long):LiveData<List<Text>>
+
+    @Query("DELETE FROM Text WHERE whiteboardId=(:whiteboardId)")
+    fun delete(whiteboardId: Long)
+
+    @Query("DELETE FROM Text WHERE id=(:textId)")
+    fun deleteText(textId:Long)
 
     @Insert
     fun insert(vararg text:Text)
 
     @Insert
-    fun insert(whiteboard:Whiteboard):Long
+    fun insert(whiteboard:WhiteboardEntity):Long
 
     @Update
     fun update(text:Text)
 
-    @Query("DELETE FROM Text WHERE whiteboardId=(:WhiteboardId)")
-    fun delete(WhiteboardId:Long)
 
-    @Query("DELETE FROM Text WHERE id=(:textId)")
-    fun delete(textId:Long)
 }
