--- WhiteboardFormatter/app/src/main/java/com/example/whiteboardformatter/data/dao/TextDao.kt
@@ -1,8 +1,0 @@
-package com.example.whiteboardformatter.data.dao
-
-import androidx.room.Dao
-
-@Dao
-interface TextDao{
-
-}
