+++ WhiteboardFormatter/app/src/main/java/com/example/whiteboardformatter/data/model/Text.kt
@@ -1,8 +1,24 @@
 package com.example.whiteboardformatter.data.model
 
 import androidx.room.Entity
+import androidx.room.ForeignKey
+import androidx.room.PrimaryKey
 
-@Entity
-data class Text(
-    var id :String
+@Entity(
+    foreignKeys = [ForeignKey(
+        entity = WhiteboardEntity::class,
+        parentColumns = arrayOf("id"),
+        childColumns = arrayOf("whiteboardId"),
+        onDelete = ForeignKey.CASCADE
+    )]
+)
+data class Text constructor(
+    @PrimaryKey(autoGenerate = true)
+    var id: Long = 0,
+    var whiteboardId: Long,
+    var text: String = "",
+    var x: Int,
+    var y: Int,
+    var width: Int,
+    var height: Int
 )
