--- WhiteboardFormatter/app/src/main/java/com/example/whiteboardformatter/data/model/Whiteboard.kt
@@ -1,8 +1,0 @@
-package com.example.whiteboardformatter.data.model
-
-import androidx.room.Entity
-
-@Entity
-data class Whiteboard(
-    var id:String
-)
